+++ src/hawkengine/src/main/java/net/hawkengine/core/materialupdater/IMaterialUpdater.java
@@ -1,0 +1,9 @@
+package net.hawkengine.core.materialupdater;
+
+import net.hawkengine.model.MaterialDefinition;
+
+public interface IMaterialUpdater {
+    MaterialDefinition getLatestMaterialVersion(MaterialDefinition materialDefinition);
+
+
+}
