+++ src/hawkengine/src/main/java/net/hawkengine/core/materialupdater/MaterialTrackerService.java
@@ -1,15 +1,13 @@
 package net.hawkengine.core.materialupdater;
 
+import net.hawkengine.model.Material;
 import net.hawkengine.model.Pipeline;
 import net.hawkengine.model.PipelineDefinition;
 import org.apache.commons.lang.text.StrBuilder;
 import org.apache.log4j.Logger;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class MaterialUpdaterService implements IMaterialUpdaterService{
-    private static final Logger LOGGER = Logger.getLogger(MaterialUpdaterService.class.getName());
+public class MaterialTrackerService implements IMaterialTrackerService {
+    private static final Logger LOGGER = Logger.getLogger(MaterialTrackerService.class.getName());
 
 
     @Override
@@ -23,7 +21,9 @@ public class MaterialUpdaterService implements IMaterialUpdaterService{
 
     @Override
     public Pipeline updatePipelineMaterials(Pipeline pipeline) {
+        for (Material material : pipeline.getMaterials()) {
 
+        }
 
         return pipeline;
     }
